requirements.txt
anytree==2.8.0
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
Django==3.1.7
django-bootstrap-modal-forms==2.1.0
django-crispy-forms==1.11.1
django-mathfilters==1.0.0
django-widget-tweaks==1.4.8
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
moviepy==1.0.3
numpy==1.20.2
Pillow==8.1.2
proglog==0.1.9
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
schedule==1.1.0
six==1.15.0
sqlparse==0.4.1
tqdm==4.60.0
urllib3==1.26.4
